Add commons for shell views
